Implement settrans --interactive
Hello,

this patch adds a new option to settrans: --interactive.  The main use
case is to be able to start a translator under gdb like so:

  settrans -i /tmp/foo gdb /hurd/hello

This is extremely useful if the translator crashes before the startup
handshake, but also just plain more convenient than using --pause.

Basically, all this feature does is start a regular command using a
regular fork and exec, with the tiny exception of making it inherit a
bootstrap port so that it can attach a translator to the specified node
(/tmp/foo in the example).  Because it is started regularly it will be
given stdin and stdout, and any other stuff that is normally inherited.

Looking closer at the implementation, you'll see that I service the
startup request in a separate thread.  This is mainly so it will work
while I use wait() to see when the child process exits.  The alternative
was to wait for a no-senders notification, but this doesn't seem robust.

As an extra, the service thread can even handle any number of requests,
which is useful if you want to restart the translator during your
debugging session.

Regards,
  Fredrik

* utils/settrans.c (options): Add `--interactive'.
(settrans_fn_t): New typedef.
(start_interactive_translator_service)
(fork_interactive_translator)
(exec_interactive_translator): New functions.
(main): Implement new `--interactive' option.