libmachdev: drop static libmachdev.so hack
Previously, netdde was built against `libmachdev.a' to avoid some
uninvestigated linking problems.  netdde using a statically linked
libmachdev lead to problems in the past.  As the linking problems only
affect functions required for block devices, which we do not use at
this point, we can simply stub them out until a proper solution is
found.

* libmachdev/Makefile: Do not overwrite the built `libmachdev.so' with
simple linker script pulling in `libmachdev.a'.